cognitionx-backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-google-genai
python-dotenv
reportlab
qrcode
pillow
cryptography
google-genai
requests

